Tolerate floating-point duration in fpcalc output
Fixes https://github.com/beetbox/beets/issues/2317

Related to https://github.com/acoustid/chromaprint/issues/30

+++ acoustid.py
@@ -300,7 +300,7 @@ def _fingerprint_file_fpcalc(path, maxlength):
             raise FingerprintGenerationError("malformed fpcalc output")
         if parts[0] == b'DURATION':
             try:
-                duration = int(parts[1])
+                duration = float(parts[1])
             except ValueError:
                 raise FingerprintGenerationError("fpcalc duration not numeric")
         elif parts[0] == b'FINGERPRINT':
